+++ packages/on_audio_query_ios/ios/on_audio_query_ios/Sources/on_audio_query_ios/PluginProvider.swift
@@ -1,3 +1,5 @@
+import Flutter
+
 /**
  * A singleton used to define all variables/methods that will be used on all plugin.
  *